Move filesystem utils and transactions into dedicated files. Additional code cleanup/refactoring.
