+++ deltacat/storage/model/table.py
@@ -11,7 +11,8 @@ from deltacat.storage.model.namespace import (
     NamespaceLocator,
     Namespace,
 )
-from deltacat.storage.model.metafile import Metafile, MetafileRevisionInfo, TXN_DIR_NAME
+from deltacat.storage.model.metafile import Metafile, MetafileRevisionInfo
+from deltacat.constants import TXN_DIR_NAME
 
 TableProperties = Dict[str, Any]
